Generic Color V2 (#48)
Improves Generic colors on browsers with support of ``color-contrast()`` (Picking the best ones, not all 18 may be seen on all possible theme combinations). It also changes message color to purple from black in accordance to this change (This applies to all browsers). Lastly, it makes an even faster experience

Browsers that do not support ``color-contrast()`` will get two sets of generic colors color-locked (One for light and one for dark themes)